refactor: register header actions only from static modules
