+++ packages/core/src/hooks/use-header-registers.js
@@ -17,5 +17,7 @@
  */
 
 export const useHeaderRegisters = headerRegisters => {
-  [...headerRegisters].map(_moduleHook => _moduleHook());
+  headerRegisters.forEach(_moduleHook => {
+    _moduleHook();
+  });
 };
